[#93] 배치 트랜잭션 수정
- FeedBatchService 클래스 생성 후 트랜잭션 적용
